// src/notifications/entities/notification-type.enum.ts
export enum NotificationType {
  ANNOUNCEMENT = 'ANNOUNCEMENT',
  LINEUP_SUBMITTED = 'LINEUP_SUBMITTED',
  LINEUP_APPROVED = 'LINEUP_APPROVED',
  LINEUP_REJECTED = 'LINEUP_REJECTED',
  LINEUP_CHANGES_REQUESTED = 'LINEUP_CHANGES_REQUESTED',
  LINEUP_MEMBER_REMOVED = 'LINEUP_MEMBER_REMOVED',
  SUBSTITUTION_REQUEST = 'SUBSTITUTION_REQUEST',
  SUBSTITUTION_APPROVED = 'SUBSTITUTION_APPROVED',
  SUBSTITUTION_REJECTED = 'SUBSTITUTION_REJECTED',
  USER_APPROVED = 'USER_APPROVED',
  GENERAL = 'GENERAL',
  PRAYER_REQUEST_PUBLISHED = 'PRAYER_REQUEST_PUBLISHED',
  PRAYER_REQUEST_PENDING = 'PRAYER_REQUEST_PENDING',
  PROFILE_CHANGE_REQUESTED = 'PROFILE_CHANGE_REQUESTED',
  PROFILE_CHANGE_APPROVED = 'PROFILE_CHANGE_APPROVED',
  PROFILE_CHANGE_REJECTED = 'PROFILE_CHANGE_REJECTED',
  LINEUP_DELETED = 'LINEUP_DELETED',
  USER_DECLINED = 'USER_DECLINED',
  ARTICLE = 'ARTICLE',
}